requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.10.5
litellm==1.76.1
tavily-python==0.3.5
httpx==0.27.0